Making GeoModel optional and moving some files
