+++ CMakeLists.txt
@@ -44,6 +44,7 @@ option(ACTS_BUILD_FATRAS_GEANT4 "Build Geant4 Fatras package" OFF)
 # examples related options
 option(ACTS_BUILD_EXAMPLES "Build standalone examples" OFF)
 option(ACTS_BUILD_EXAMPLES_DD4HEP "Build DD4hep-based code in the examples" OFF)
+option(ACTS_BUILD_EXAMPLES_GEOMODEL "Build GeoModel-based code in the examples" OFF)
 option(ACTS_BUILD_EXAMPLES_GEANT4 "Build Geant4-based code in the examples" OFF)
 option(ACTS_BUILD_EXAMPLES_HEPMC3 "Build HepMC3-based code in the examples" OFF)
 option(ACTS_BUILD_EXAMPLES_PYTHIA8 "Build Pythia8-based code in the examples" OFF)
@@ -222,6 +223,10 @@ endif()
 if(ACTS_BUILD_EXAMPLES_GEANT4)
   find_package(Geant4 REQUIRED CONFIG COMPONENTS gdml)
 endif()
+if(ACTS_BUILD_EXAMPLES_GEOMODEL)
+  find_package( GeoModelCore PATHS /afs/cern.ch/work/f/fklimpel/install/ REQUIRED ${GeoModel_VERSION} )
+  find_package( GeoModelG4 PATHS /afs/cern.ch/work/f/fklimpel/install/ REQUIRED ${GeoModel_VERSION} )
+endif()
 if(ACTS_BUILD_EXAMPLES_HEPMC3)
   find_package(HepMC3 ${_acts_hepmc3_version} REQUIRED CONFIG)
 endif()
@@ -234,8 +239,6 @@ if(ACTS_BUILD_DOCS)
   find_package(Sphinx REQUIRED)
 endif()
 
-find_package( GeoModelCore PATHS /afs/cern.ch/work/f/fklimpel/install/ REQUIRED ${GeoModel_VERSION} )
-find_package( GeoModelG4 PATHS /afs/cern.ch/work/f/fklimpel/install/ REQUIRED ${GeoModel_VERSION} )
 
 if(ACTS_CUSTOM_SCALARTYPE)
     message(STATUS "Building Acts with custom scalar type: ${ACTS_CUSTOM_SCALARTYPE}")
