+++ Examples/Algorithms/Geant4HepMC/src/EventRecording.cpp
@@ -10,7 +10,6 @@
 
 #include "ActsExamples/EventData/SimParticle.hpp"
 #include "ActsExamples/Framework/WhiteBoard.hpp"
-#include "ActsExamples/Geant4/GdmlDetectorConstruction.hpp"
 
 #include <iostream>
 #include <stdexcept>
