+++ Examples/Run/Geant4/CMakeLists.txt
@@ -1,21 +1,3 @@
-add_executable(
-  ActsExampleGeantinoRecordingGdml
-  GdmlGeantinoRecording.cpp)
-target_link_libraries(
-  ActsExampleGeantinoRecordingGdml
-  PRIVATE
-    ActsExamplesCommon
-    ActsExamplesDetectorDD4hep
-    ActsExamplesFramework
-    ActsExamplesGeant4
-    ActsExamplesGeant4GeoModel
-    ActsExamplesIoRoot
-    Boost::program_options)
-
-install(
-  TARGETS 
-    ActsExampleGeantinoRecordingGdml 
-  RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR})
-
 add_subdirectory_if(DD4hep ACTS_BUILD_EXAMPLES_DD4HEP)
 add_subdirectory_if(HepMC ACTS_BUILD_EXAMPLES_DD4HEP AND ACTS_BUILD_EXAMPLES_HEPMC3)
+add_subdirectory_if(GeoModel ACTS_BUILD_EXAMPLES_GEOMODEL)
