+++ Examples/Run/Geant4/GeoModel/GdmlGeantinoRecording.cpp
@@ -6,12 +6,12 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "ActsExamples/Geant4/GdmlDetectorConstruction.hpp"
+#include "ActsExamples/Geant4GeoModel/GdmlDetectorConstruction.hpp"
 #include "ActsExamples/Options/CommonOptions.hpp"
 
 #include <boost/program_options.hpp>
 
-#include "GeantinoRecordingBase.hpp"
+#include "../GeantinoRecordingBase.hpp"
 
 using namespace ActsExamples;
 using namespace ActsExamples;
